requirements.txt
Django==4.0.5
djangorestframework==3.13.1
keras==2.9.0
numpy==1.23.0
pandas==1.4.3
sklearn==0.0
tensorflow==2.9.1
tensorflow-estimator==2.9.0
scikit-learn>=1.0.0
scikeras==0.8.0
django-environ==0.9.0
tokenizers==0.12.1
xgboost==1.6.1
bs4==0.0.1
python-whois==0.8.0
lxml==4.9.1
tqdm==4.64.0